+++ CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- [store-deno-kv] & [store-deno-kv-fs] renamed `getKv()` to `getDenoKv()`
+
+### Added
+
+- [store-deno-kv/get-deno-kv] utility function `getDenoKv()` to obtain `Deno.Kv`
+  from a store if it exposes a `getDenoKv()` function
+
+### Fixed
+
+- stricter TS options, add `import type` where necessary
+
 ## [0.2.0]
 
 ### Changed